requirements.txt

Flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
rapidfuzz==3.9.7
gunicorn==22.0.0
Jinja2==3.1.4
MarkupSafe==2.1.3
Werkzeug==3.0.3
itsdangerous==2.2.0
click==8.1.7
blinker==1.8.2
openai==1.51.0
anthropic==0.34.2
httpx==0.27.2
